Make schema optional for migration statements
MySQL and MariaDB define databases and schemas as synonyms. Forcing
users to define a schema would not make any sense if MySQL or MariaDB
support is added in the future. Also if only one schema is being used in
all migrations, forcing the user to explicitly state the schema for
every table will become very repetitive. By making the schema optional,
it is possible to automatically use the default schema if no schema has
been defined or to define a schema for the whole migration.